requirements.txt
Flask
python-dotenv
requests
flask_session